chore: move convert to errors to internal directory
